olive_disease_app/requirements.txt
# ── Web Servers ──────────────────────────────────────────────────────────────
flask>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0

# ── Machine Learning ──────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
# YOLOv5 (via torch.hub — pas besoin de l'installer manuellement)
# ou si vous avez cloné YOLOv5 localement :
# -r yolov5/requirements.txt

# ── Image processing ─────────────────────────────────────────────────────────
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pandas>=2.0.0

# ── Utils ────────────────────────────────────────────────────────────────────
pydantic>=2.0.0
python-dotenv>=1.0.0
